Add more rally benchmarks related to Cinder
Some of Rally benchmarks related to Cinder are failing often.
Especially benchmarks that attach volumes.

Adding all Cinder related benchmarks to cinder gate to help
Cinder team to fix issues.

As well make this job put to Jenkins +1/-1.
For that we added for every benchmark sla section
that checks that there is no failed iterations.

Change-Id: I3b32dd91792445d26755e8d469edb5d3f82ebca2