VMware: Create volume backing in specific clusters
The VMDK driver doesn't allow specifying a set of vCenter
clusters as the target for volume backing creation unlike
the VMware Nova driver. This patch adds support for an
optional list of vCenter cluster names in Cinder conf which
will be used by the VMDK driver as targets for volume backing
creation.

DocImpact
    Added a new config option 'vmware_cluster_name' which
    specifies a vCenter compute cluster where volumes
    should be created.

Change-Id: I0dcb3a8ac7c9eaa0d0697f4967873d82bf1bbddf