VMware: Change inventory folder hierarchy
In vCenter inventory, volumes are organized under the folder:
<datacenter_vm_folder>/<volume_folder>, where <volume_folder>
is the value of driver config option 'vmware_volume_folder'.
Due to this organization certain project-specific maintenance
tasks (using vSphere client) might be difficult for vCenter
admin. For example, the admin may need to migrate the volumes
of a particular project from one vCenter datastore to another.

This patch changes the folder hierarchy to:
  <datacenter_vm_folder>/OpenStack/Project
                               (<project_id>)/<volume_folder>
for better organization.

DocImpact
  Modified default value of config option
  'vmware_volume_folder' and updated its description.

Closes-Bug: #1485550
Change-Id: I37a3bd00efaa3b4a0bbcb22cbe9fff3f4fb80456