Adds pool aware scheduling for HNAS drivers
This commit adds the pool aware feature for HNAS drivers. Now, each service
label in HNAS XML config will be considered as a pool unit. In order to be able
to cast a volume to a service, the admin must create 'volume_type's and each
volume type need to have the 'service_label' with the same tag set in the XML
config, and then the user can create a volume with the desired volume_type.

Implements:  blueprint hds-hnas-pool-aware-sched
Change-Id:  Iddff7ad606b06454458fc53bfc7b812f9815f0e1