Use lockutils.set_defaults to set lock_path in test
Commit 9dd6d212f0d8fd74c29077a786fb6d770a2a0bba for oslo.concurrency's
lockutils appears to have broken our unit tests.

Twelve of the the storwize_svc unit tests started failing with the
complaint that the required configuration option 'lock_path' was not set.

This change updates cinder/test.py to not only set up a temporary
lock_path location that will be automatically cleaned up, but to also
register that temporary lock_path location using lockutils.set_defaults()
to correct the problem.

Change-Id: I011888f676ebc170e7285048b7e5a638a5a8c9a9
Closes-bug: 1410485