HP3Par: Set snapCPG when managing existing volumes
This fixes a corner case where the user did not assign a "Copy CPG"
value to the volume when originally created on the back-end, and
then the user does not assign a "Volume Type" to the volume
when it is then "managed" into OpenStack.

This results in an undefined "snapCPG" value which means no
snapshots can be created for the volume.

The fix is to set the "snapCPG" value to match the "User CPG"
value associated with the volume when the volume is "managed"

Change-Id: I207d462e12e4db67bff8624bc69e7e6f1ceff75b
Closes-Bug: #1393609