Switch the PureISCSIDriver over to using the purestorage pypi module.
All management api's will now go through the purestorage module, the
interface is extremely similar to the original FlashArray object
included in the cinder.volume.drivers.pure module. This allows for
updates to the REST API and python code that uses it to not disrupt the
cinder driver.

Implements: blueprint pure-iscsi-purestorage-lib
Change-Id: Ifcfb1c2337e5837f2121929c4788137d8bd1cf42