Limit volume copy bandwidth per backend
Currently, by setting volume_copy_bps_limit in cinder.conf,
volume/image copy can be throttled to mitigate slow down of
instances' volume access during the copy. However, this is a global
setting and applied to all the backends.
This change enables admins to configure volume_copy_bps_limit to
different values among backends by specifying it in each backend
section.

In addition, with this change, the bps limit will be divided when
multiple volume copy operations run concurrently on the same backend.
For example, if volume_copy_bps_limit is set to 100MB/s, and 2 copies
are running on a backend, each copy can use up to 50MB/s.
This behavior will be useful for QoS which ensures a certain amount
of bandwidth is kept for instances' volume access.

This introduces Throttle classes to count volume copies in a backend
and setup bandwidth limit.

Change-Id: Ie390b46538556fa704b06ffc79cd6cc000bd5936
Implents: blueprint limit-volume-copy-bps-per-backend