Fix the eqlx driver to retry on ssh timeout
When the ssh session is timing out, the driver
should make attempts to retry based on the value
in eqlx_cli_max_retries. Instead it was raising
the exception and bailing out on a single attempt.

Fixed the driver to raise a different exception
so the ssh sessions can be retried.

Added unit tests to ensure the max retries
happen

Also fixed the actual attempts made in the message

Change-Id: I9bda46f9ef63eec436a381da3c96c09ce77c31ef
Closes-Bug: #1412940