Change volume and snapshot stuck creating to error
Service of volume goes down after the status of
the volume is turned 'creating', volume to remain with
the status of the 'creating'.

Status of the volume is still 'creating' even if
volume service is restarted, it is impossible for user
to remove it without force-delete.

With this patch, in order to keep consistence with Nova,
volume state 'creating' should be shifted to the 'error' properly.

Closes-Bug: #1310526
Co-Authored-By: Ivan Kolodyazhny <e0ne@e0ne.info>
Change-Id: I89ef201f25b739cc443cb2920de2bd6ce89d97e1