Add retry to lvm snapshot create
We have some occasional issues with snapshot-create
failing for what looks to be conflicts with udev.  It
looks like this problem is a status conflict between LVM
cache and udev, and in most cases the best way to get
around this is to retry the command a few times until
the cache and udev are back in sync.

This patch uses the newly added retry decorator and
for now we're just adding it to the snapshot create
call.  We're using the default values for interval and
retry count but we can certainly adjust this as needed.

Change-Id: Ic884b829591361da979fde968f5710237b34565f
Closes-Bug: #1335905