Remove usage of taskflow 'utils.misc' module
The failure type at its old location in an internal
utils directory of taskflow is deprecated (in general
usage of taskflow utils code should be restricted/not
done) and it has been moved to a public location of
taskflow.types (which is ok to use and is not
deprecated).

This change updates to use the better and more supported
module/code location instead.

Change-Id: I8b8c6e98c934fbf1341b8f1c907ecd385405a3c1