HP 3PAR modules have bad log messages
Some log messages are incorrectly formatted. Specifically, those
messages that take multiple variables. In this case, each variable
needs to be assigned a unique placeholder.

This defect was recently introduced when all log messages where
updated to use "," instead of "%" (see
Change-Id: I23d53d66fda47981ed8f20b618b2ced8ef6e0682),
which allowed translations to be ignored if not required by current
log levels.

Change-Id: Ic3439ba24c61579457dee206064bed2465ae4b6b
Closes-Bug: #1411370