Add debug message for lvremove after udev settle
Just add a simple debug level log message for lvremove calls that
are made and successfully completed as part of error recovery.

Currently some platforms commonly receive a "Unable to deactivate"
error message when running an lvremove.  We have a retry mechanism
that performs a udevsettle and retries that in most cases results
in a succesful removal.

This patch just adds an explicit debug log message so we can track
this more easily and gather some statistical data on it.

Change-Id: I8f2480491f3cb7f078597113f967c1e285fdf105