Add completion logging for snapshots and volumes
While trying to debug insufficient resource issues occasionally
seen in the gate, it was noted that there's no logging for
resource deletion of volumes and snapshots.

This patch adds some simple logging in the cinder.volume.api
to note when and why a delete call might fail as an info
message indicating when the request is successfully issued
to the scheduler/rpc layer.

In addition, this patch also adds the same sort of logging
to the LVM backend driver.

This additional logging will help troubleshoot the current
resource issues/races we're encountering in the gate.

Change-Id: I1ac39a43d63650ab020b180eefb3a24a946328d2