Remove the import of oslo_i18n lib in Juno
To address bug 1408099 I submitted a patch that
removed the calls to the private method in the i18n
code (change I2ae3d9b98c107cebaf386adbdcdb3cfafee070be).

Unfortunately as can be seen in the patch review,
my work was modified by another contributor and an update
of the oslo_18n lib which included a test fixture to togle
lazy translations was added.

This was probably fine for Master, HOWEVER that patch was
then back ported to Juno, which was completely inappropriate
as Juno was not using the oslo_i18n lib but using
openstack/common/gettextutils (change 894f20d).

This patch modifies the change that was submitted and removes the
import of the oslo_i18n lib which should have never been introduced
in stable in the first place.

Before somebody proposes it, leaving the lib dependency and capping
it is NOT the right answer here.  The lib should absolutely not be
introduced/back-ported in to stable.  Especially in this case where
it simply is used for a unit test that quite frankly is a crap test
that shouldn't be there to begin with.  There is no reason I can see
at all why we are testing translations in Cinder unit tests, that
should be left to the unit test for the library itself.

Change-Id: I5776d9f54d7a85d74311d9b8dbbe28fb81c62027
Closes-Bug: #1420335