VMware:Use datastore selection logic in new module
Volumes created by the VMDK driver are stored in vCenter
managed datastores. A datastore is selected based on
requirements such as size, storage policy etc. Commit
2c672d1100ad4f44517838e17156b3ea6300b1cc introduced a
new module for datastore selection. It improved the
current logic and grouped the code which were scattered
in multiple places (in vmdk.py). This patch refactors
vmdk module to remove duplicate code and use the methods
in the new module.

Change-Id: I3b7cb8bd33ced2f9deed2936a9c02feb90fb70b7