Tests: Remove TestWSGIService.test_reset_pool_size_to_default
This test invokes Cinder WSGIService.Service() and asserts
against values that are only manipulated in wsgi.Server(),
which WSGIService is based on.

We also have TestWSGIServer.test_reset_pool_size_to_default
which does the same thing for wsgi.Server().

Remove this test since it duplicates the TestWSGIServer test
and tests code which is an additional layer of abstraction
(cinder.WSGIService) away from the class that the code belongs
to (wsgi.Server).

Change-Id: I68d6f479da26e75618bfa773c84273501926ab07