Add project_id to barbican keymgr wrapper
Currently, the barbican keymgr wrapper does not provide the project_id
to barbican.  As a result, all key operations done in cinder with a
barbican backend are stored in the barbican database with a NULL
external project_id.

This patch adds the project_id to the auth plugin created using the
auth token, so that barbican has access to the project_id and uses
the external project_id with all key operations.

Change-Id: Ia5f6cdd5177893d2f13f181e1ced278d49f9d910
Closes-Bug: #1421795