Fix return value inconsistency in VNX Driver
When iscsi_initiators option for EMC Cinder Driver has no value,
a dict object is returned by find_available_iscsi_targets() in
EMC Cinder Driver while a list is expected. Thus an
initialize_connection() will fail due to the inconsistency.

This patch is to fix this inconsistency in the driver.

Change-Id: I094340cc60da4a8b423233983a1d00784bbbb8ea
Closes-Bug: 1422563