Sync 'eventlet_backdoor' module from oslo-incubator
The eventlet_backdor module hasn't been updated in Cinder since
September of 2014.  In order to be able to move to the oslo.log
library and be able o clean other modules out of oslo-incubator
we need to bring it up to the latest level.

Current HEAD in OSLO:
---------------------
commit e589dde0721a0a67e4030813e582afec6e70d042
Date:  Wed Feb 18 03:08:12 2015 +0000
Merge "Have a little fun with release notes"

Changes merged with this patch:
---------------------
2aacb111 - Change oslo.config to oslo_config
2fbf5065 - Remove oslo.log code and clean up versionutils API
80c47519 - Remove unnecessary import of eventlet
a3220c51 - add list_opts to all modules with configuration options
5d40e143 - Remove code that moved to oslo.i18n

Change-Id: I875964a00e449e2314a648cc9586d97e51990de8