Sync 'threadgroup' from oslo-incubator
The threadgroup module hasn't had a sync since September of 2014.
The module needs to be updated to support the move to using the
oslo_log library.  This patch brings us up to the latest module
level.

Current HEAD in OSLO:
---------------------
commit e589dde0721a0a67e4030813e582afec6e70d042
Date:  Wed Feb 18 03:08:12 2015 +0000
Merge "Have a little fun with release notes"

Changes merged with this patch:
---------------------
2fbf5065 - Remove oslo.log code and clean up versionutils API
1523f000 - threadgroup: don't log GreenletExit

Change-Id: I5b7e281f6879f2749cd8b446c65e2fcb3abfbe95