Sync 'versionutils' module from oslo-incubator
The versionutils module hasn't had a sync since November of 2014.
There have been namespace changes and changes to enable moving to
the new oslo_log library.  So, we need this sync so that we can
move forward with those changes.

Current HEAD in OSLO:
---------------------
commit e589dde0721a0a67e4030813e582afec6e70d042
Date:  Wed Feb 18 03:08:12 2015 +0000
Merge "Have a little fun with release notes"

Changes merged with this patch:
---------------------
9bf01f9d - Switch from oslo.config to oslo_config
2fbf5065 - Remove oslo.log code and clean up versionutils API

Change-Id: I4344ed8fb234135a199649c5ef7620e4166fbb12