Dell Storage Center Driver API bug
In later versions of the Dell Storage Center REST API
the index attribute is removed from the ScVolume object.
Referencing this caused an exception that failed volume
creation.  Replaced with instanceId.

Also _init_volume was mapping/unmapping to every server
that was currently up on the Storage Center rather than
a single server.  While this would certainly initialize
the volume it would create a lot of extra work on a busy
system.  We now break after the first map/unmap.

Changed copyright to 2015 to match when it will actually
be released.

Change-Id: I89f3967144d7effbb67d1498fc3b4a7a23042a59
Closes-Bug: 1426463