Tests: Harden fake_notifier asserts
This is a continuation of
8740e063 - Make test_create_delete_snapshot more robust

The same issue cropped up for me in
test_create_consistencygroup_from_src().

Add some more specific asserts to try to reveal what
is going wrong here.  The idea is to assert more
specific things (examining individual notifications)
before more general asserts (count of notifications) when
possible, to fail the test on the check that will provide
the most information back.

Related-Bug: #1412513

Change-Id: I267bd5933247f69f220a7be871a63e4db9eb9a80