Create unit tests for volume objects
This is a follow up patch to cinder objects that adds
missing unit tests for volume objects. It also adds
a @base.remotable decorator to the volume save method,
which makes sure a context is properly supplied to the
method, whether it is passed in or taken from the
context stashed in the object.

Change-Id: I9febfdf94c4d8c82defedf0ceaec03800552affe
Implements: blueprint cinder-objects
Closes-Bug: #1434571