Make the 3PAR drivers honor the pool in create
Currently, the 3PAR drivers rely on the pool (CPG) existing
in the extra specs of a volume type.   This is the way that the
3PAR drivers supported the idea of 'pools' prior to Cinder
itself supporting pools.

Now, the 3PAR drivers will not use the CPG specified in the
extra specs of a volume type.   The drivers will log a warning
explaining that CPG is deprecated, and then override the CPG
setting with the pool name specified by the scheduler in the
volume['host'] entry.    This fixes a few bugs in the 3PAR
drivers.

Existing 3PAR deployments will need to update their cinder.conf
to specify the list of CPGs they want the driver to support as
pools in the hp3par_cpg=[list of cpgs] entry.  hp3par_cpg can be
a single CPG, or a comma separated list of CPGS.

The 3PAR best practice guide will be updated for the Kilo release
explaining how to upgrade from existing deployments and describe
how best to use CPGS as pools.

Closes-Bug: #1432876
DocImpact

Change-Id: I2c733ff3f13ba309cc07e07331b4945b1981a07a