Fix remaining memory issues with nfs backup unit tests
Here we fix another place in the nfs backup unit tests where
the configured file size was larger than the chunk size in
the corresponding Swift backup tests.  I missed this one in

   https://review.openstack.org/#/c/165394

and even though code with that fix ran fine in a bare 1GB VM,
there are still memory issues when run in CI with a VM of that
size.

Closes-Bug: 1452811
Change-Id: I20dcfb657699e1b415ae2e36ec63f8c6b0049285