Set volume_attachment to [] for the temporary volume creation
A new table volume_attachment has been added recently. During the
creation of the temporary volume in migration, the attribute
volume_attachment is not empty if an attached volume is going to be
migrated, because all the properties are simply copied.
This non-empty attribute volume_attachment leads to the failure of
the volume creation. This patch removes this attribute before the
temporary volume creation.

Besides, detach the volume should consider multi-attach as well by
adding the attachment id parameter.

Change-Id: I7c2f69945ee8d72b2dba4e68a91b872207df5310
Closes-bug: #1433019