Fix incorrect invocation of _add_to_threadpool
_add_to_threadpool expects to be passed the function and which is to be
run in an alternate thread, and the function's arguments. Instead, the
function (self.delete_volume) was being invoked directly, and it's
result passed to _add_to_threadpool. This results in a TypeError being
raised (the result is a bool, which is not callable), but more
importantly, it defeats the purpose of the statement - to offload the
invocation of delete_volume to another thread.

Change-Id: I524156cbffe0f6f5a2da902c6a62417d1ec425b9
Closes-Bug: #1436624