Removed sleep before 'YES' is sent to confim an operation
Removed sleep between the time a command is sent
and 'YES' is sent to confim the operation. We have
discussed this internally and have concluded that
the sleep was mistakenly added in the initial coding
from an example of the SSH code. We have done full
functionality testing of the zoning operations on
Brocade switches with the sleep removed and have
found no side effects from the removal

Additionally, no other Brocade management
applications which perform zoning on our switches
require sleep intervals between execution of
commands. We are confident that the sleep is not
necessary and can be removed

Change-Id: I2ac6e50ff3d231b4d404f1a0be29f907e8d20458
Closes-Bug: #1436594