Reworked Dell SC iSCSI target portal return
On initialize_connection the code to determine the portal, lun and
iqn info to return could skip ports and potentially not return the
best portal choice.  In the case of multipath being enabled not all
ports would be returned.

Also changed a LOG.debug to a LOG.info in initialize_connection.

Closes-bug: 1442346

Change-Id: I2c5235da4d373d15303325a51ed1d4c37fd38815