VolMgr: reschedule only when filter_properties has retry
In the task flow for volume manager, create_volume tasks, volume gets
reschedule even when scheduler doesn't indicate so.  The problem is
the flow should not only check 'allow_reschedule' and 'request_specs',
but also (more importantly) filter_properties['retry'], which is
populated by scheduler if schedule_max_attempts is set to > 1.  This
checks was there before taskflow was introduced, but somehow the
migration missed the check for filter_properties['retry'].

This change adds back the check, so scheduler_max_attempts won't be
treated like scheduler_max_attempts = infinite.

Change-Id: Ia873617b22a2d86662ea8ea2fec5aae7f54a2058
Closes-bug: #1445561