Add resource tag to logging in volume.manager.py
We now have resource tag support in oslo logging,
and our logging is pretty inconsistent and down right
ugly in places.  Let's clean things up based on the
standard logging guidelines and use the fancy new
resource tag.

To use set the following in cinder.conf:
logging_context_format_string = \
%(asctime)s.%(msecs)03d %(levelname)s %(name)s [%(request_id)s \
%(project_name)s] %(resource)s%(message)s

This change hits the majority of the code in manager and
should be used as an example/guide.  There are some exceptions
around migration and replication where things are kinda ugly,
those should be fixed up as follow up work.

Change-Id: I1315cc16116601fea8c023b47f99b2eaad01d6fb