Add support for volume groups and netapp_raid_type
This patch removes a filter constraint on pool information that limited the
pools that would be returned to the scheduler based on the raid type for
the DDP feature of the E-Series platform. Now that pools may have different
RAID types, we return that as an unqualified extra-spec netapp_raid_type.

This patch also resolves a defect with the reporting of E-Series
extra-specs that was introduced with the removal of the cache functionality.

DocImpact
Partially-Implements blueprint netapp-eseries-additional-extra-specs
Change-Id: I2656c3282270a5ec5256cc12767a09138c4a5371