Release Import of Translations from Transifex
This updates also recreates all pot (translation source files) to
reflect the state of the repository.

This change needs to be done manually since the automatic import does
not handle the proposed branches and we need to sync with latest
translations.

Note: This is part of importing of translations, there are no new
translations for this project, thus only the pot file gets updated.

Change-Id: Ib472c156a7007b2166c6857ae1fa2bb72a2e0c1a