Create initiator id for VMAX iSCSI driver
This is related to Bug #1440427. The fix for that bug only addressed
the issue for the FC driver. The same problem exists for the iSCSI
driver. This patch addressed the problem for the VMAX iSCSI driver.
The initiator id will be created on the array if it does not exist
for the iSCSI driver.

Change-Id: I5df57df122a704066d4d7a1997ac2642b073c9e8
Closes-Bug: #1446443