service child process normal SIGTERM exit
service.py had some code where the child process would catch the
SIGTERM from the parent just so it could exit with 1 status rather
than with an indication that it exited due to SIGTERM. When
shutting down the parent doesn't care in what way the child ended,
only that they're all gone, so this code is unnecessary.

Also, for some reason this caused the child to never exit while
there was an open connection from a client. Probably something
with eventlet and signal handling.

This is a cherry-pick of oslo-incubator commit
702bc569987854b602ef189655c201c348de84cb .

Change-Id: I87f3ca4da64fb8070e4d6c3876a2f1ce1a3ca71d
Closes-Bug: #1446583