Fix namespace issue in generate_sample.sh
We just moved to using a hybrid approach that
uses the new oslo-config-generator for pulling out
library options to create the cinder.sample.conf file.
In that patch it was expected that we should use oslo_*
for the namespace.  That, however, is not how
oslo-config-generator works.

This patch switches to using oslo.* for the namespace to get
the library options pulled in.  This also fixes the fact that
we were putting multiple '[DEFAULT]' sections into the file.

Change-Id: Ie799d4d7e6deb2c486a93c5095fc3b766679fa95
Partial-bug:  1449235