Dell: Added support for update_migrated_volume
Added support for updated_migrated_volume to common driver.
Added rename_volume to dell_storagecenter_api to support it.
Added tests to test_dellsc.py and test_dellapi.py.

The only thing we do in this rename the newly migrated volume
destination volume on the Dell Storage center to be the same
as the original volume ID.  Since we identify the volumes
by volume ID this prevents them from being lost.

Change-Id: I531e24853cf5db8b7c34d91460d3abe786a652de
Closes-bug: 1452919