3PAR enable multiattach capability reporting
This patch enables the reporting of the
multiattach capability in get_volume_stats.
The scheduler currently filters on the capability
for drivers that support/don't support that ability
to do more than one attachment for a volume.
By reporting multiattach = True, it lets the scheduler
know this driver allows attaching a volume more than once.

Change-Id: Ia18414bbadff14777c2266b681c03c76a2aa6889