Add secondary account capability to SolidFire
The SolidFire Cluster has a limit of 2K volumes per SolidFire
account. Given that the SolidFire account maps directly to the
OpenStack tenant this can be a limitation for some large scale
deployments that may have tenants that wish to create more than
2K volumes in their OpenStack cloud.

This patch introduces the concept of a secondary account for
an OpenStack tenant on the SolidFire backend.

If an OpenStack tenant has more than 2K volumes, and wishes to
create additional volumes, we can create a secondary/supplemental
SolidFire account that is owned by the same OpenStack tenant.

If they wish to go beyond 4K volumes... well, SolidFire V3
driver will deal with that, for now 4K is a line in the sand.

While working on unit tests, we again move to using mock
opportunistically.  In addition, a few tests that either
weren't actually testing anything, or were just a fake-loop
that didn't actually work have been removed:

NOTE:
This patch is a temporary for backporting purposes, we are
still planning a refresh in Liberty of the driver.

Change-Id: I1c8a14e6bea1b6f47d10777c56c5af42b63e2dae