Replace dit.itervalues() with dict.values()
This change adds Python 3 compatibility to the modified code.

The itervalues() method of Python 2 dictionaries was renamed to values() on
Python 3. As discussed on the openstack-dev mailing list, itervalues() must be
replaced with values(), six.itervalues() should not be used. In OpenStack, the
overhead of creating a temporary list with dict.values() on Python 2 is
negligible.

Blueprint cinder-python3

Change-Id: Ibcf4597aeef1835f08bee2d6831205c428884c31