Get updated volume status in begin_detaching
Heat stack deletion followed by an unsuccessful heat stack delete
attempt resulted in calling begin_detaching on a volume which was
already in detaching state. It was observed that the API returned
success instead of failing the operation. There is a check which
verifies the volume status, but it is not using the updated volume
status. This patch fixes the problem by getting updated volume
status from DB. But this fix  won't eliminate the possibility of
above mentioned inconsistent behavior if multiple begin_detaching
happens concurrently.

Closes-Bug: #1466060
Change-Id: I2a4cacef193e6cfa68026d97292370a895b291a9