Fix broken export commands on block_device driver
When I implemented the new target driver code the
methods used for LVM were copy pasted:
 commit 9651f547147188645942466602c92cce06666483

I screwed this up because the block_device driver
doesn't use LVM at all.

This patch sets it back the way it should be and just
uses the local dev file of the whole disk.

Change-Id: Ife88d5597df27dcd668a9a361bebccbc840ca2d4
Closes-Bug: #1465819