Port NetApp E-Series iSCSI driver to ABC model
This commit modifies the NetApp E-Series volume driver to the
new ABC driver model.  The corresponding FibreChannel driver
already uses that model since it was recently implemented and
was written with an appropriate ABC class signature in the first
place.  A corresponding class signature is now used for iSCSI.

This commit also modifies a couple doc strings to match the
corresponding FibreChannel driver.

Change-Id: I624efd6e511f15d34fff1d97a0848c2610185e17
Partially-implements: blueprint abc-driver-update