Return 404 if volume type encryption is not found
This commit adds a check for volume type encryption existence before
marking it as deleted. If it isn't present call returns 404 error
code so that it can be handled properly in services that use the
REST API (like Heat).

APIImpact
HTTP 404 will be returned if volume type encryption is not found.

Change-Id: I3107b8df0121de1ebe9e812ef7a842d09500e030
Closes-Bug: #1468751