Handle attachment of second volume
When the zoning policy is initiator and a second volume is attached
to an instance, the current implementation deletes the existing zones
and then adds new zone data. While doing so, the configuration is deleted
and it is never added back.

The issue is due to delete/add design. When zones are deleted, active zone
set should also be updated, as based on active zone set config is either
created/updated. Added code to get active zone set before adding the zones.

Change-Id: Ib0cd51c3724142175eb1c602d6c1f87dd5e985fb
Closes-Bug: #1465207