Tests: test_volume mock conversion
The goal here is to try to drive down complexity in
our tests.  Note that this file is massive and this
will be a gradual effort, not something that will
be completed in one patchset.

osprofiler is mocked out in all unit tests now, so
remove that from here.

Make a first pass at converting these tests over to
mock and strengthen the assertions for calling methods
that were previously mocked with stubs.

Remove a bunch of "fake" methods, using mock instead.

Switch to mock decorators for some tests where it
makes the code easier to follow.  This helps because
it separates test setup from the test itself more
clearly.

Change-Id: Ifdf2980d183cf83c88848a20519403648427a644