Mock socket.gethostbyaddr in test_v7000_fcp
Currently test_v7000_fcp doesn't mock socket.gethostbyaddr and
hardcodes expected response to be google-public-dns-a.google.com
(as it is querying for 8.8.8.8 as IP). This introduces dependency on
external DNS lookup to pass unit tests and causes problems with
running tests behind a proxy.

This commit mocks socket.gethostbyaddr in the problematic test.

Change-Id: I3a0288d8e51c832a72e16e23af5adbee990f9e20
Closes-Bug: 1470113