Move update_migrated_volume() to BaseVD
The update_migrated_volume() function on volume/manager.py
calls the update_migrated_volume() from the driver. If the
driver does not implement the function, the manager falls
back (by catching the NotImplementedError exception) and
implements a default behavior, not renaming the volume.  If
the driver does not have the function implemented, the
migration will fail as in this case, an AttributeError
exception can be raised.

We fix this by implementing the method
update_migrate_volume() in the BaseVD driver so all drivers
now throw the proper exception.

Change-Id: I9f3cbdc5ae1cdcbf8fe61168abf35b985d4182c6
Closes-Bug: #1471807