Remove unused arguments from c-vol's create_volume
Arguments snapshot_id, image_id, source_volid, source_replicaid,
consistencygroup_id and cgsnapshot_id in volume.rpcapi.create_volume
RPC method are duplicated either in request_spec dictionary passed as
one of the arguments or are saved into the databased and are retrieved
by flow processing the request.

To simplify the flow this commit removes these duplicated parameters
and adapts rest of the code to use occurrences that are already there.

Related-Blueprint: taskflow-refactoring
Change-Id: I4dbb57968358e8930b923275c2dfd0e44d7b40d8