Fix block_device driver to behave as documented
According to the original blueprint
(https://blueprints.launchpad.net/cinder/+spec/block-device-driver) the
intended behavior of the "BlockDeviceDriver" was to directly attach the block
device to the instances using virtio (without LVM and iSCSI involved) when the
volume's underlying device is located on the same host as the instance that it
is going to be attached to. This is also how it worked through Juno. It got
broken during the Kilo cycle as part of this change:
https://review.openstack.org/#/c/135139/
It seems that the change in behavior was not intended by that patch as it
addressed something completely different. And it was neither mentioned in the
commit message nor anywhere in the review comments.

This patch restores the original behavior.

Change-Id: Ic22f03860a43b767e9097ce6dc5b397cf9f3f076
Closes-Bug: 1473001