Adding delete-wait-loop for CloudByte Volumes
Deleting volumes in CloudByte Storage is an asynchronous process.
So, the driver needs to wait till the job/process is completed.
The driver uses the jobid in a wait loop for a configured interval
of time, till the volume deletion is completed.

Added two new config options for this operation:

"cb_confirm_volume_delete_retry_interval"
"cb_confirm_volume_delete_retries"

DocImpact

Change-Id: I8139ecfb1b49c0b69ed956e3a53b5bb895ffc1fa