Prevent missing Purity hosts from raising errors
Occasionally a Purity host can be deleted out from underneath Cinder.
When this happens if cinder is attempting to disconnect a volume from
that host we should treat it as a ‘success’ since if the host is gone,
it is disconnected already (as far as Purity is concerned).

To prevent one cause of this, the disconnect method will now be locked
with the connect ones so that can’t delete a host we intend to re-use.

Change-Id: Iba82c93f9e101709cf5532225068474688187b29
Closes-bug: #1472710