Dell SC: Fix Consistency Group issues
The Dell Storage Center driver was not honoring the consisgroup_id
in the volume object in volume creation. Updated create_volume,
create_cloned_volume and create_volume_from_snapshot to honor this
value.

The driver was also not checking that volumes were initialized prior
to snapshotting a consistency group. The Dell backend does not
actually create the volumes until there is data or the volumes are
mapped to a server.

On create_volume_from_snapshot the driver was not checking for
cgsnapshot_id before snapshot_id.

Changed from calling the db to update snapshots to retrieving
the SnapshotList() object.

In addition some tests were not checking results. Fixed that.

Change-Id: I896f7438c0b98003aa345ca1c3841aba8835cc2a
Closes-Bug: #1474062
Closes-Bug: #1474105
Closes-Bug: #1475016