Dell SC: Fix legacy bug, init_conn bug and REST API bug.
If Dell iSCSI ports are in legacy mode rather than virtual port
mode initialize_connection would fail to return connection information.

If initialize_connection was called more than once for the same
connection it would attempt to map the volumes again and fail. It should
instead just return the connection information and not try to map the
volumes again.

The Dell REST API 2.2 and later changed the format of the payload
filter.  We need to support either format.

Finally consistency group support is only supported in Dell REST API
2.1 or later.  If the version is 2.0 we raise NotImplemented.

Change-Id: Ie798c49a591abb510bd947fc89c26de18f0f3c87
Closes-Bug: #1476786
Closes-Bug: #1476794
Closes-Bug: #1476797