NetApp ESeries: fix delete of non-existent volume
If a non existing volume is deleted the NetApp
iSCSI eseries driver throws a KeyError Exception
in the delete_volume method.
This situation can occur if the user manually deletes
a LUN in the NetApp controller directly or if the
delete operation initiated via cinder is interrupted,
which leaves the system in an inconsistent state.
For example if the volume is deleted on the NetApp
controller but still residing in the OpenStack database.

Change-Id: I4e9388c620556ed042b7e3fe59b550871df5b563
Closes-Bug: #1477027