Implement retype for Pure drivers
Currently retyping on the same backend requires a migration to occur
due to using the default implementation of retype on the base volume
driver.

Volumes on the Pure arrays do not have any differentiation
between volume types Cinder may be using. Because of this we can
implement a retype method that will be used whenever changing
types on the same backend that essentially does nothing except return.

Change-Id: Ibf4c57166fe5742cc7e079c4422f7efd2b6dce79
Closes-Bug: 1477780