Start/Stop coordinator with Volume service
To make use of the coordinator we need to start and stop it together
with the service that uses it.

This patch adds start and stop method calls in Volume service and
makes it easier to start it from other services.

Restored from https://review.openstack.org/#/c/205839

Co-Authored-By: Gorka Eguileor <geguileo@redhat.com>
Co-Authored-By: Szymon Borkowski <szymon.borkowski@intel.com>
Implements: blueprint cinder-volume-active-active-support
Change-Id: Ie5c551f3c7441eb7dc55fa62bed3db8c5a97e7fd