Fix volume limit exceeded exception
The VolumeLimitExceeded exception contains the wrong limit when
the number of volumes exceeds the quota corresponding to a volume
type. It always contains the limit of quota 'volumes'. This patch
fixes the limit in both exception and log message. It also adds
the quota name to exception and log message for better clarity.

Change-Id: I3e2d4fa00fe49f64b2153930e901fccd8015483e
Closes-Bug: #1481182