Support host type specific block volume attachment
----
NOTE(markus_z): This is a "backport" of commit
b1a8c2e1f18d424adfdfa29bae47cafa4700dcee which is merged in os-brick.
That patch is marked as implementing a blueprint, but should have
been a bug in the first place. The blueprint support-hosttype-on-attach
shouldn't have been necessary.
----

Add the platform and operating system type to the connector to allow a
Cinder driver to set host type specific configuration attributes when
attaching a Fibre-Channel volume.

Closes-Bug: #1471810

Change-Id: I2eef20b46808a49a330c64d517ad6430392674fd