Fix the virtual port support in VNX driver
VNX support virtual port on each physical iSCSI port. The
driver used some incorrect command which cannot get the virtual
port id information. This causes the driver to not be able to
tell whether the registered port is a virtaul port or not.
This change fixes the driver to use correct command and handle
the virtual port id correctly.

Change-Id: I11994dd5a0c8da74fcf15668f49796a45719b5e4
Closes-Bug: #1485484