Assisted volume migration for Oracle ZFSSA drivers
The Oracle ZFSSA drivers will migrate the volume directly from the
source host to the destination host if both hosts are configured
with Oracle ZFSSA. This type of volume migration will bypass the
cinder host and will directly transfer the volume to the destination
backend.

The ZFSSA iSCSI driver will use zfssa's remote replication service
to move the the volume to the destination zfssa.

Implements: blueprint oracle-zfssa-volume-migration
Change-Id: I4de8108e3c6c0969c83e87eac139f04b4fe8c7ac