Allow CG without snapshot to be deleted
If a CG is created from CG Snapshot, it cannot be deleted until
the CG Snapshot is deleted. This is wrong. This is because the query
to find CG who are parents of CG Snapshot didn't filter out the
CG uuid and as a result, it is not possible to delete any CG if
there are any existing CG Snanpshot.

This patch fixes the problem by fixing the function in db API
that finds all CGs that have CG Snapshot depending on them.
* CGs with any CG Snapshot depending on them still cannot be deleted
  until the CG Snapshot is deleted first.
* CGs without any CG Snapshot depending on them should be allowed
  to be deleted.

Change-Id: I1380b26c4a957946c7dc27285557966f558cf03d
Closes-Bug: #1485783