Prevent volume already in CG to be added to another
If a volume is already in a CG, currently it can still be added to
another CG without any warning. We should add a check to see if the
volume is already in a CG and prevent it from being added to another
if that check returns True. It has to be removed from the previous
CG in order to be added to a new one.

Change-Id: I6ec92cbf4ce60e7afbfa5a57edef18ddba554a74
Closes-Bug: #1487151