Fix QoS keys not being available to scheduler
There is a qos_specs property contained within the properties
passed to filters and weighers during volume placement decision
time. Currently qos_specs is always set to None even when a
QoS spec is associated with a volume type.

This patch exposes the QoS keys by placing them in the
qos_specs property. When no QoS spec is associated with a volume
the qos_specs property would still be set to None.

Change-Id: I5d5d36b56efeadd0884abc2f541ccf44b1010a2e
Closes-Bug: #1486286