Add ability to update provider_id during init
During the Liberty release we added the ability to store
a provider_id for a volume in the Cinder database.  This
simplifies a number of things for back ends when accessing
volumes.

One missing piece here however is that we have no way to
populate this data for existing deployments if/when they
upgrade.

This patch provides a simple solution that issues an update
call to the drivers on host init.  It works by fetching a
list of volumes for the specific host, and passes those in
to the driver.  The driver can do nothing, or it can go
through and sync up the Cinder ID's with it's internal
ID's and return the provider update info.

Change-Id: Ib3b078a6f492afada24e534f380f5f014033d603