Update provider_id column on SolidFire init
This patch just implements an init routine to update
the cinder db with provider_id mappings to known/active
volumes on the SolidFire cluster.

Now that we have added a mapping feature in the DB, we need
to go back and populate these columns on an upgrade.  It's
also helpful to have this mechanism for internal mappings on
init, that's why it's here as an init routine instead of part
of an upgrade migration or cinder-manager script.

Change-Id: I1e4da2adff2ae85a3020b38167c8c63fc56b8417