CG driver function should not access db
Currently the following CG driver interfaces don't pass in volumes
or snapshots to the driver. As a result, drivers have to retrieve info
from db. In this patch, these interfaces are changed so that drivers
don't have to access db.

Removing the access to db from driver CG functions will be handled by
separate patches.

* create_consistencygroup
* delete_consistencygroup
* create_cgsnapshot
* delete_cgsnapshot

Change-Id: I63287fd43927fc09dad77292ecfe561917f7bda5
Closes-Bug: #1501466