Dell SC: init_volume stale volume info fix
In init_volume the volume information becomes stale after the volume
is mapped to a server.  The volume should be updated before being sent
to unmap.

The unmap function was testing for the volume being active.  This
should not be required.  It is unlikely an inactive volume will have
mappings but not impossible.  This check has been removed.

Removed LOG.error from _find_mapping_profiles as the API returning
an error is fine if the volume is inactive.  Plus we were not failing
on it.  The LOG.debug is sufficient.

Change-Id: I3ff8447360cea134e1008fa8db5cd912aa5effd3