3PAR Disable generic image volume cache
This patch disables the 3PAR drivers if an admin tries to
enable the generic image volume cache capability.

The 3PAR drivers don't work with this currently because of the way volume
cloning works in the drivers.  When a volume is cloned, it's done so
in the background, which makes the clone available immediately for attaching.
The problem is, a background cloning also locks the original volume from
being modified.  The generic image cache does a clone, and then an extend
volume, which fails for the 3PAR drivers currently.

Change-Id: If2e79a24969ab85bacdffdf750af5fc74be37a5c
Partial-Bug: 1491088