Use consolidated update for failover_replication
Right now it’s using a special built volume_updates dictionary that
contains some fields *and* a model update. We then glom all of them
together into an update dictionary for the volume model.

This change simplifies them into a single model update more in-line with
how the other volume API’s work for drivers.

It also fixes a bug where the model_update property of the volume_update
was not being applied to the db model.

Change-Id: I8ad85bf0e62099a8ea4dd471dfcb64fa057f1021
Closes-Bug: #1491587