3PAR fix driver to work with image cache
This patch fixes the problem with the 3PAR not
working with the generic image cache.

In order to do the image cache, the 3PAR has to create a
temporary snapshot of the volume being cloned.  Then the
background cloning is against the temporary snapshot.
The problem is, the temporary snapshot stays around.
We try to account for this temporary snapshot in
the delete volume action.

We also bump the minimum required python-3parclient
version to 4.1.0, which has the new getVolumeSnapshots
API call required to make this all work.

Change-Id: I6dea01ce07387990aa38bd5106b51504739b3a6e
Closes-Bug: #1491088