Don't rely on provider_id for resource deletion
Provider_id info isn't necessarily reliable when dealing
with failure cases.  There are a number of cases where the
info in the Cinder database may not be updated.

This change uses the old delete routines instead of trying
to use the provider_id info.  There are ways to do both,
however it's probably not that beneficial for now.

Change-Id: Idbfd98b19f3c7f1b231ed26a2cc6c98aa2899e2a
Closes-Bug: #1494597