CG creation should be scheduled on backend level
Currently, cinder scheduler schedules the CG creation
on pool level, that means if there are multiple volume
types in one CG if each volume type is associated with a
different pool (those pools under same backend), you can't
create the CG successfully.

According to the design of CG's, it should be scheduled on
backend level to support the case mentioned above.

DocImpact
UpgradeImpact
Change-Id: I335dbdd1d2b1146543148ccce8d52162804369f4
Closes-Bug: #1496655