NetApp: Fix issue with updating E-Series password
The NetApp E-Series storage array password is specified in the
configuration file, and must be updated if the array has its password
changed via the management software. However, the logic to check
whether or not the password is currently valid was incorrect, and
was not correctly detecting an invalid password in need of updating.

This patch fixes the password validity checking for the driver,
refactors the existing code to be more concise, and adds additional
unittests around this functionality.

Change-Id: I1468d1810479947a9cff7ef8b79440de4063cbed
Closes-Bug: #1497007